Add tray support to the application
This commit adds simple tray support to Ongaku. Clicking the close
button or the tray icon will hide the application. If hidden clicking
the tray icon will restore the application. There is also a show/hide
toggle menu option. The application can be exited via the quit menu item.

This also adds templates for an app icon as one was needed to show the
tray icon.

Fixes #12